+++ core/src/main/java/com/yahoo/ycsb/generator/UniformIntegerGenerator.java
@@ -22,11 +22,11 @@ import java.util.Random;
 import com.yahoo.ycsb.Utils;
 
 /**
- * Generates integers randomly uniform from an interval.
+ * Generates longs randomly uniform from an interval.
  */
 public class UniformIntegerGenerator extends IntegerGenerator 
 {
-	int _lb,_ub,_interval;
+	long _lb,_ub,_interval;
 	
 	/**
 	 * Creates a generator that will return integers uniformly randomly from the interval [lb,ub] inclusive (that is, lb and ub are possible values)
@@ -34,7 +34,7 @@ public class UniformIntegerGenerator extends IntegerGenerator
 	 * @param lb the lower bound (inclusive) of generated values
 	 * @param ub the upper bound (inclusive) of generated values
 	 */
-	public UniformIntegerGenerator(int lb, int ub)
+	public UniformIntegerGenerator(long lb, long ub)
 	{
 		_lb=lb;
 		_ub=ub;
@@ -42,9 +42,9 @@ public class UniformIntegerGenerator extends IntegerGenerator
 	}
 	
 	@Override
-	public int nextInt() 
+	public long nextInt() 
 	{
-		int ret=Utils.random().nextInt(_interval)+_lb;
+		long ret=Utils.random().nextInt((int)_interval)+_lb;
 		setLastInt(ret);
 		
 		return ret;
